requirements.txt
pandas
numpy
plotnine
plotly
tabulate
pytorch
coral_pytorch
tqdm
tabulate
scipy
sklearn
mtat
